Added Zoom
Added zoom on viewport coordinates; need to refactor so that the Zoom
ops don't have any exposed variables to the MainForm (obviously), but
I'll take a holistic look at that later.

Added a debug form, but I need to play with data types in Pascal more
before I go off half-cocked and implement something terrible.

Still working on it, but it's enough to commit for now! Also need to
decide how to netlist the schematic items for external simulators.